Add parens around converse implication example
I think this is right, and I think it increases readability. Not being familiar with the operator, I didn't know the precedence, and so didn't know how to parse the example.

Alternatively, I think you could eliminate the `== bool` term, since the next line gives the value of the expression.